Generalized Test cases for dump()
